Refactor: controller에 plan패키지 -> travelplan으로 변경
